feat: implement modules and hash tasks
